+++ alpine.dockerfile
@@ -35,11 +35,13 @@ RUN pip wheel -r tests/requirements-tools.txt -w wheels
 
 # Recent versions of python docker image do not provide setuptools and wheel with python (>= 3.12) by default.
 # See: https://github.com/docker-library/python/issues/952
-RUN pip install --upgrade setuptools build
+RUN pip install --upgrade hatchling
 
 # Build a wheel for PyInstaller. Do this last and use as few files as possible to maximize cache-ability.
 COPY COPYING.txt .
-COPY setup.* ./
+COPY README.rst .
+COPY hatch_build.py .
+COPY pyproject.toml .
 COPY bootloader bootloader
 COPY PyInstaller PyInstaller
 RUN pip wheel --no-build-isolation --no-dependencies --wheel-dir=wheels .
@@ -59,7 +61,7 @@ RUN apk add libpng
 # And by PyInstaller itself.
 RUN apk add binutils
 
-COPY setup.cfg .
+COPY pyproject.toml pytest.ini .
 
 # Import and the precompiled wheels from the `build` image.
 COPY --from=wheel-factory /io/wheels /wheels
