--- setup.cfg
@@ -1,121 +1,0 @@
-[metadata]
-name = pyinstaller
-version = attr: PyInstaller.__version__
-url = https://www.pyinstaller.org/
-project_urls =
-    Source = https://github.com/pyinstaller/pyinstaller
-
-description = PyInstaller bundles a Python application and all its dependencies into a single package.
-# Long description consists of README.rst only
-long_description = file: README.rst
-long_description_content_type = text/x-rst
-
-author = Hartmut Goebel, Giovanni Bajo, David Vierra, David Cortesi, Martin Zibricky
-
-
-keywords =
-    packaging, app, apps, bundle, convert, standalone, executable
-    pyinstaller, cxfreeze, freeze, py2exe, py2app, bbfreeze
-
-license = GPLv2-or-later with a special exception which allows to use PyInstaller to build and distribute non-free programs (including commercial ones)
-license_files = COPYING.txt
-
-classifiers =
-    Development Status :: 6 - Mature
-    Environment :: Console
-    Intended Audience :: Developers
-    Intended Audience :: Other Audience
-    Intended Audience :: System Administrators
-    License :: OSI Approved :: GNU General Public License v2 (GPLv2)
-    Natural Language :: English
-    Operating System :: MacOS :: MacOS X
-    Operating System :: Microsoft :: Windows
-    Operating System :: POSIX
-    Operating System :: POSIX :: AIX
-    Operating System :: POSIX :: BSD
-    Operating System :: POSIX :: Linux
-    Operating System :: POSIX :: SunOS/Solaris
-    Programming Language :: C
-    Programming Language :: Python
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-    Programming Language :: Python :: 3.12
-    Programming Language :: Python :: 3.13
-    Programming Language :: Python :: 3.14
-    Programming Language :: Python :: Implementation :: CPython
-    Topic :: Software Development
-    Topic :: Software Development :: Build Tools
-    Topic :: Software Development :: Interpreters
-    Topic :: Software Development :: Libraries :: Python Modules
-    Topic :: System :: Installation/Setup
-    Topic :: System :: Software Distribution
-    Topic :: Utilities
-
-[options]
-packages = find:
-zip_safe = False
-include_package_data = False
-python_requires = >=3.8, <3.15
-install_requires =
-    setuptools >= 42.0.0
-    altgraph
-    # pefile 2024.8.26 contains performance regression that heavily impacts our binary dependency analysis
-    pefile >= 2022.5.30, != 2024.8.26; sys_platform == 'win32'
-    pywin32-ctypes >= 0.2.1 ; sys_platform == 'win32'
-    macholib >= 1.8 ; sys_platform == 'darwin'
-    pyinstaller-hooks-contrib >= 2025.8
-    importlib_metadata >= 4.6 ; python_version < "3.10"
-    packaging >= 22.0
-
-[options.packages.find]
-include =
-    PyInstaller
-    PyInstaller.*
-
-[options.package_data]
-PyInstaller =
-    # Include all bootloaders in wheels by default.
-    bootloader/*/*
-    # These files need to be explicitly included as well.
-    fake-modules/*.py
-    fake-modules/_pyi_rth_utils/*.py
-    hooks/rthooks.dat
-    lib/README.rst
-
-[options.extras_require]
-; for 3rd-party packages testing their hooks in their CI:
-hook_testing =
-    pytest >= 2.7.3
-    execnet >= 1.5.0
-    psutil
-; for CLI tab completion
-completion =
-    argcomplete
-
-[options.entry_points]
-console_scripts =
-    pyinstaller = PyInstaller.__main__:_console_script_run
-    pyi-archive_viewer = PyInstaller.utils.cliutils.archive_viewer:run
-    pyi-bindepend = PyInstaller.utils.cliutils.bindepend:run
-    pyi-grab_version = PyInstaller.utils.cliutils.grab_version:run
-    pyi-makespec = PyInstaller.utils.cliutils.makespec:run
-    pyi-set_version = PyInstaller.utils.cliutils.set_version:run
-
-[sdist]
-# For release distribution generate .tar.gz archives only. These are
-# about 10% smaller then .zip files.
-formats=gztar
-
-#[bdist_wheel]
-# We MUST NOT create an universal wheel as PyInstaller has different
-# dependencies per platforms and version and includes compiled binaries.
-#universal = MUST NOT
-
-[clean]
-# Always fully clean. Otherwise, bootloaders for one platform remain in the
-# build cache and end up blindly copied into wheels for another platform.
-all=1
